// frontend/src/api/generated/models/alertHistoryResponse.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * Chain Risk Platform - BFF
 * BFF API for Chain Risk Platform
 * OpenAPI spec version: 1.0
 */
import type { AlertHistoryResponseSeverity } from './alertHistoryResponseSeverity';
import type { AlertHistoryResponseMetadata } from './alertHistoryResponseMetadata';
import type { AlertHistoryResponseStatus } from './alertHistoryResponseStatus';

export interface AlertHistoryResponse {
  id: number;
  ruleId?: number;
  alertType: string;
  severity: AlertHistoryResponseSeverity;
  entityType: string;
  entityId: string;
  title: string;
  message: string;
  metadata: AlertHistoryResponseMetadata;
  status: AlertHistoryResponseStatus;
  notifiedAt?: string;
  acknowledgedAt?: string;
  acknowledgedBy?: string;
  createdAt: string;
}